+++ src/main/java/jp/aquafactory/apprenticecodex/compat/malum/MalumHauntedCompat.java
@@ -7,9 +7,9 @@ import jp.aquafactory.apprenticecodex.item.AbstractRightClickMagicWeaponItem;
 import jp.aquafactory.apprenticecodex.item.chargedtwinbladestaff.ChargedTwinBladeStaff;
 import jp.aquafactory.apprenticecodex.item.focusstaffbow.FocusStaffbow;
 import jp.aquafactory.apprenticecodex.item.manaforceblade.ManaForceBlade;
-import jp.aquafactory.apprenticecodex.item.MulticastEchoStaff;
-import jp.aquafactory.apprenticecodex.item.PastelStaff;
-import jp.aquafactory.apprenticecodex.item.ZenithStaff;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaff;
+import jp.aquafactory.apprenticecodex.item.pastelstaff.PastelStaff;
+import jp.aquafactory.apprenticecodex.item.zenithstaff.ZenithStaff;
 import jp.aquafactory.apprenticecodex.utility.CombatTools;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.damagesource.DamageSource;
