+++ src/main/java/jp/aquafactory/apprenticecodex/item/multicastechostaff/MulticastEchoStaff.java
@@ -1,57 +1,57 @@
-package jp.aquafactory.apprenticecodex.item;
+package jp.aquafactory.apprenticecodex.item.multicastechostaff;
 
 import io.redspace.ironsspellbooks.api.registry.AttributeRegistry;
+import io.redspace.ironsspellbooks.api.spells.IPresetSpellContainer;
+import io.redspace.ironsspellbooks.api.spells.ISpellContainer;
 import io.redspace.ironsspellbooks.item.UniqueItem;
 import io.redspace.ironsspellbooks.item.weapons.AttributeContainer;
 import io.redspace.ironsspellbooks.item.weapons.StaffItem;
 import io.redspace.ironsspellbooks.item.weapons.StaffTier;
 import io.redspace.ironsspellbooks.render.ClientStaffItemExtensions;
-import jp.aquafactory.apprenticecodex.item.zenithstaff.ZenithStaffClientTooltip;
-import jp.aquafactory.apprenticecodex.renderer.item.ZenithStaffRenderer;
+import jp.aquafactory.apprenticecodex.item.StaffEnchantmentTargeting;
+import jp.aquafactory.apprenticecodex.registry.SpellRegistry;
+import jp.aquafactory.apprenticecodex.renderer.item.MulticastEchoStaffRenderer;
+import jp.aquafactory.apprenticecodex.utility.InitialSpellContainerHelper;
 import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Rarity;
-import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
-import net.minecraft.world.level.Level;
-import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.extensions.common.IClientItemExtensions;
-import net.minecraftforge.fml.DistExecutor;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import software.bernie.geckolib.animatable.GeoItem;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
-import software.bernie.geckolib.core.animation.AnimationController;
 import software.bernie.geckolib.core.animation.AnimatableManager;
-import software.bernie.geckolib.core.animation.RawAnimation;
-import software.bernie.geckolib.core.object.PlayState;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
-import java.util.List;
-import java.util.Optional;
 import java.util.function.Consumer;
 
-public class ZenithStaff extends StaffItem implements GeoItem, UniqueItem {
-    private static final String MAIN_CONTROLLER = "main";
-    private static final RawAnimation ANIM_IDLE = RawAnimation.begin().thenLoop("idle");
-    private static final StaffTier ZENITH_STAFF_TIER = new StaffTier(
+public class MulticastEchoStaff extends StaffItem implements GeoItem, IPresetSpellContainer, UniqueItem {
+    private static final StaffTier MULTICAST_ECHO_STAFF_TIER = new StaffTier(
             3.0F,
             -3.0F,
             new AttributeContainer(
                     AttributeRegistry.SPELL_POWER,
-                    0.10D,
+                    0.05D,
+                    AttributeModifier.Operation.MULTIPLY_BASE
+            ),
+            new AttributeContainer(
+                    AttributeRegistry.ELDRITCH_SPELL_POWER,
+                    0.15D,
+                    AttributeModifier.Operation.MULTIPLY_BASE
+            ),
+            new AttributeContainer(
+                    AttributeRegistry.CAST_TIME_REDUCTION,
+                    0.2D,
                     AttributeModifier.Operation.MULTIPLY_BASE
             )
     );
 
     private final AnimatableInstanceCache cache = GeckoLibUtil.createInstanceCache(this);
 
-    public ZenithStaff() {
-        super(new Item.Properties().stacksTo(1).rarity(Rarity.EPIC).fireResistant(), ZENITH_STAFF_TIER);
+    public MulticastEchoStaff() {
+        super(new Item.Properties().stacksTo(1).rarity(Rarity.EPIC).fireResistant(), MULTICAST_ECHO_STAFF_TIER);
         GeoItem.registerSyncedAnimatable(this);
     }
 
@@ -63,12 +63,12 @@ public class ZenithStaff extends StaffItem implements GeoItem, UniqueItem {
     @Override
     public void initializeClient(Consumer<IClientItemExtensions> consumer) {
         consumer.accept(new ClientStaffItemExtensions() {
-            private ZenithStaffRenderer renderer;
+            private MulticastEchoStaffRenderer renderer;
 
             @Override
             public BlockEntityWithoutLevelRenderer getCustomRenderer() {
                 if (renderer == null) {
-                    renderer = new ZenithStaffRenderer();
+                    renderer = new MulticastEchoStaffRenderer();
                 }
 
                 return renderer;
@@ -78,12 +78,15 @@ public class ZenithStaff extends StaffItem implements GeoItem, UniqueItem {
 
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllerRegistrar) {
-        controllerRegistrar.add(
-                new AnimationController<>(this, MAIN_CONTROLLER, 0, state -> {
-                    state.setAnimation(ANIM_IDLE);
-                    return PlayState.CONTINUE;
-                })
-        );
+    }
+
+    @Override
+    public void initializeSpellContainer(ItemStack itemStack) {
+        if (itemStack == null || itemStack.isEmpty() || ISpellContainer.isSpellContainer(itemStack)) {
+            return;
+        }
+
+        InitialSpellContainerHelper.setInitialContainer(itemStack, 1, true, false, SpellRegistry.ECHO_CAST, 1);
     }
 
     @Override
@@ -103,7 +106,11 @@ public class ZenithStaff extends StaffItem implements GeoItem, UniqueItem {
         }
 
         var enchantments = EnchantmentHelper.getEnchantments(book);
-        return enchantments.isEmpty() || enchantments.keySet().stream()
+        if (enchantments.isEmpty()) {
+            return true;
+        }
+
+        return enchantments.keySet().stream()
                 .allMatch(enchantment -> canApplyAtEnchantingTable(stack, enchantment));
     }
 
@@ -113,25 +120,7 @@ public class ZenithStaff extends StaffItem implements GeoItem, UniqueItem {
         return 22;
     }
 
-    @Override
-    public void appendHoverText(
-            @NotNull ItemStack itemStack,
-            @Nullable Level level,
-            @NotNull List<Component> lines,
-            @NotNull TooltipFlag flag
-    ) {
-        super.appendHoverText(itemStack, level, lines, flag);
-
-        var tooltipLines = resolveClientTooltipLines();
-        tooltipLines.ifPresent(lines::addAll);
-    }
-
-    public static boolean isZenithStaff(ItemStack stack) {
-        return stack != null && !stack.isEmpty() && stack.getItem() instanceof ZenithStaff;
-    }
-
-    private static Optional<List<Component>> resolveClientTooltipLines() {
-        var result = DistExecutor.unsafeCallWhenOn(Dist.CLIENT, () -> ZenithStaffClientTooltip::createLines);
-        return Optional.ofNullable(result);
+    public static boolean isMulticastEchoStaff(ItemStack stack) {
+        return stack != null && !stack.isEmpty() && stack.getItem() instanceof MulticastEchoStaff;
     }
 }
