+++ src/main/java/jp/aquafactory/apprenticecodex/item/multicastechostaff/MulticastEchoStaffCastHelper.java
@@ -10,7 +10,6 @@ import io.redspace.ironsspellbooks.api.spells.ICastData;
 import io.redspace.ironsspellbooks.config.ServerConfigs;
 import jp.aquafactory.apprenticecodex.ApprenticeCodex;
 import jp.aquafactory.apprenticecodex.config.ApprenticeCodexServerConfig;
-import jp.aquafactory.apprenticecodex.item.MulticastEchoStaff;
 import jp.aquafactory.apprenticecodex.registry.EffectRegistry;
 import jp.aquafactory.apprenticecodex.registry.SpellRegistry;
 import jp.aquafactory.apprenticecodex.utility.SpellCooldownHelper;
