+++ src/main/java/jp/aquafactory/apprenticecodex/item/multicastechostaff/MulticastEchoStaffClientRenderState.java
@@ -2,7 +2,6 @@ package jp.aquafactory.apprenticecodex.item.multicastechostaff;
 
 import io.redspace.ironsspellbooks.api.magic.SpellSelectionManager;
 import io.redspace.ironsspellbooks.player.ClientMagicData;
-import jp.aquafactory.apprenticecodex.item.MulticastEchoStaff;
 import jp.aquafactory.apprenticecodex.registry.EffectRegistry;
 import jp.aquafactory.apprenticecodex.registry.SpellRegistry;
 import net.minecraft.client.Minecraft;
