+++ src/main/java/jp/aquafactory/apprenticecodex/item/zenithstaff/ZenithStaffClientRenderState.java
@@ -1,7 +1,6 @@
 package jp.aquafactory.apprenticecodex.item.zenithstaff;
 
 import io.redspace.ironsspellbooks.api.spells.SchoolType;
-import jp.aquafactory.apprenticecodex.item.ZenithStaff;
 import jp.aquafactory.apprenticecodex.utility.MagicTools;
 import net.minecraft.client.Minecraft;
 import net.minecraft.util.Mth;
